requirements.txt
psycopg2-binary==2.8.4
dj_database_url==0.5.0
certifi==2020.6.20
chardet==3.0.4
Django==2.2.13
djangorestframework==3.12.1
idna==2.10
pytz==2020.4
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
